test(dsh-provider-usage): unit-chart 并入 unit-v1 修复双通道死文件缺口（#150 变异加固 2/4）
unit-chart.test.ts 此前未被 smoke.ts import、也不在 stryker testFiles，
其图表断言从未进入任何执行通道。整体并入 unit-v1.test.ts（smoke 与
stryker 双通道均执行）后删除死文件，图表纯函数群与 adapter format
断言正式纳入变异计量。
